fix: rename image fetcher file
